rename tx_digest to digest as most types use Wrapper(Digest) where Digest is Base58-able, so we don't need to limit to tx_digest only
