+++ crates/sui-graphql-rpc/src/context_data/sui_sdk_data_provider.rs
@@ -11,6 +11,7 @@ use crate::types::big_int::BigInt;
 use crate::types::checkpoint::Checkpoint;
 use crate::types::committee_member::CommitteeMember;
 use crate::types::date_time::DateTime;
+use crate::types::digest::TransactionDigest;
 use crate::types::end_of_epoch_data::EndOfEpochData;
 use crate::types::epoch::Epoch;
 use crate::types::object::{Object, ObjectFilter, ObjectKind};
@@ -23,7 +24,6 @@ use crate::types::storage_fund::StorageFund;
 use crate::types::sui_address::SuiAddress;
 use crate::types::system_parameters::SystemParameters;
 use crate::types::transaction_block::TransactionBlock;
-use crate::types::tx_digest::TransactionDigest;
 use crate::types::validator::Validator;
 use crate::types::validator_credentials::ValidatorCredentials;
 use crate::types::validator_set::ValidatorSet;
