+++ crates/sui-graphql-rpc/src/types/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod checkpoint;
 pub(crate) mod coin;
 pub(crate) mod committee_member;
 pub(crate) mod date_time;
+pub(crate) mod digest;
 pub(crate) mod display;
 pub(crate) mod end_of_epoch_data;
 pub(crate) mod epoch;
@@ -25,7 +26,6 @@ pub(crate) mod storage_fund;
 pub(crate) mod sui_address;
 pub(crate) mod system_parameters;
 pub(crate) mod transaction_block;
-pub(crate) mod tx_digest;
 pub(crate) mod validator;
 pub(crate) mod validator_credentials;
 pub(crate) mod validator_set;
